Better names, consolidate jobs, reorder parts
Better (imo) names. They're not installing dependencies so much as
installing system components. Yeeees these are kinda the same thing, but
"dependencies" implies (again: imo) that there is a package we're
building, and these the things we need for it to work. The script isn't
that.

Parts have been reordered, and consolidated. It doesn't seem at all
necessary to update cache, upgrade, add repo, update cache, install,
upgrade and clean. We can add the key, repo file, and then update cache,
install and upgrade (and clean), all together.