Partial guide to VirtBox usage
It's gonna take me a hot minute to figure out what to include, and what
to skip past to not confuse people. VirtualBox doesn't behave quite the
way I expected, but it's not exactly difficult. Still, I gotta make sure
that people with *zero* prior experience can get this up and running on
their own.