Finish VBox install part
Turns out the VBox unattended install process depends on files that are
optional on some (my) systems. I don't want to deal with the various
ways to get those things in what is meant to be a beginner guide.
Instead, we'll just do it "manually" and move on. On and on to the
Ubuntu setup process!