Installing Ubuntu finished, part of ROS2
Ubuntu installation is straight forward enough, but dropping new users
off so suddenly isn't a good call. Ciaran decided on a list of steps to
hit on the way through the install and first-time-setup process. We're
finally into the ROS2 setup procedures! We're collecting tools, touching
on package management, and then reaching for automation to pick up the
rest.