Mostly correct groups and hosts
Use the host names instead of the group names as the hosts for the guihosts group. Now the local connecitons work, though it does mean we have two "local" hosts. Which, uh, can't be correct. We'll need to address this somehow. The devkit could (and maybe should) just live in its own inventory file.