Finish section on running Ansible
Completed instructions for running the Ansible playbook agains the local
machine. There's a new inventory file that is only the local machine.
This is an easier way to address the existence of multiple hosts all
claiming to be "local" when they cannot possibly all be. There are some
goings-on outside this branch that I know will make this largely
irrelevant soon, but this is what I'm doing to reach the merge event.