Play and group structure based on goals
The plays operate on groups of hosts, and those hosts can all be made to
be similar in goal. There are placeholder plays (Deck) and largely
incomplete plays (Devkit), and their matching entries in the inventory.
But it's explicit this way. I may look into Ansible Roles if this
becomes too unwieldy, and I can find the time.