Syntax fix, type enforcement
List the packages in correct YAML format, you dummy.

It seems that Ansible (sometimes..?) doesn't like the "true" or "yes"
value for the upgrade-and-clean phase. It converts these to a string and
emits a warning about it. For some reason a string instead of one of the
enumerated possible values is acceptable instead of an error, but here
we are.