Several playbooks aimed at particular roles
I've separated almost all the tasks in to their own playbooks. This is,
of course, absurd, but I'm still learning how Ansible works. The idea is
to both have parts that can be composed based on roles, and to separate
the parts that call `pip` from the become root directives.