Proper Ansible-Roles structure taking shape
Some heavy layout changes, now as I get into the Ansible roles features.
It seems that there's better defined support for things like static
data, as well (regarding the apt-key usage, and the plan to replace it
with direct copying of the ros gpg key into the hosts). No idea if the
play(s) work right now, but we're getting there.