New playbook and host files
The main playbook exists again! With it, a new set of inventory files
which should be a more fitting structure. The playbook should be able to
call upon roles to install the components to each category of host, and
the inventory files should be grouping the hosts according to the
components they need.