Apply ansible-lint suggestions
Yay for quality tooling. The only thing that's going to be properly
difficult to fix is the state:latest for systemd. I think it can become
a handler, and set it up to trigger only when too old. For now, it does
what we need, and the linter is only warning us about good practice.
It's not fatal.