For pip, become the ansible_user in inventory
We'll still need to decide on what we want the user to be called, and
how it should be set up, but we can at least push that decision out of
the playbook. It's an inventory property, now. Of course, we gotta keep
those properties attached to all the hosts, now, so that'll be fun.