Add play for upgrading hosts
For whatever reason, the apt package lists won't be updated even when
they're greater than 24 hours old. This caused the real Raspberry Pi
host to try to download old packages that were no longer available on
the servers. We were also failing to upgrade anything other than the
systemd package which is pinned specifically to "latest". This corrects
a usage error (upgrade and install) from a previous commit, and adds an
explicit upgrade (and clean) step to the common role.