Fixup ROV setup.py, and supply ament marker file
Fill out (some) details of seahawk_rov/setup.py, and re-enable the
ament marker resource install. This installs a file named the same as
the package. It's source-tree location is ${package}/resource/, and
appears to need exactly nothing written in it.

(Also: Do the same for the "our_examples" package. It's not real
important, but it'll cause `colcon` to abort builds if it doesn't also
work).