Reorder declarations and assignments in IMU code
This change brings the order of all the feature enable as well as the
destructured msg assignments to bring them into order of how they are
stored within the ROS2's Imu class. Pretty much just semantic, but it
will hopefully make it easier to read. I also fixed a spelling error.
It is worth noting the order of destructuring might not be correct.