Make ESC Lerp more meaningful + ESC deadzone
Looking at BlueRobotics' data table for their T200 thruster, there are
actually only so many values under 5 amps. Ciaran already limited the
values that could possibly be sent to the ESCs with the servokit
library, but there was another issue: the -1,1 float range from
/drive/motors was mapped directly to the 0-3000 value range of the ESC,
80% of which we couldn't use at all due to the power draw being more
than 5 amps. So, we just pegged the values for 40% of our input range at
1220, and another 40% of our input range at 1780. In other words, only
20% of our range was controllable. This commit improves the accuracy by
changing the linear interpolation parameters such that we can actually
use most of the values.