The start of where I would go with it
This is the basic brush up of what I would do for a config class. I've
added a new field (hardware_location) but removed the sensor values. I
figure that the configuration could be repurposed for the
ThrustBoxBME280, and so it should know how to name it's topics. And that
the t, h, p weren't needed. They could have been the suffixes of the
topic names, but simply imposing the "/temperature", "/humidity", and
"/pressure" names is probably fine. If it conflicts, we can rewrite the
code at that time.

I'm not sure I'd have a `class Config` at all, however. These things are
easy enough to feed in as arguments, and we're not doing any kind of
dynamic reconfiguration that might be facilitated by having a class to
hold that logic.