BME280: Class float instead of type float
For some inane and unknowable reason, despite working with and storing
the float type internally, these message types try to assert whether the
value passed to them is of the float class, rather than the float type.
As such, I've put the float type values returned from the BME280 into
float classes instead. So it should be happy now and not crash
unpredictably. Why is Python like this? This sucks so much. If I didn't
have fancy syntax highlighting in VSCode I legitimately never would have
figured out the problem.