DELETE ZTRIM
ztrim was never a good idea, it should have been implemented as vertical
twist increase, but its always been broken and now it is gone. goodbye
forever.

this commit removes all references to ztrim and using lb/rb (i think??)
and also removes the 'y' button which i didn't even know was a thing and
is documented literally nowhere.