chore: Switch to repo submodule from setup folder
This begins the standardization process on the setup repository, rather
than just having the folder as part of the parent repository as we did
previously. This has the advantage of making it easier to migrate setup
changes between different years/robots more easily, and this carries
over to this repository as well with the `seahawk_1` branch.