chore: Switch h264 submodule to relative path
This allows for any protocol to be used to clone the submodule, rather
than just https. It defaults to whatever the parent was cloned with, so
if you cloned with ssh, it would clone the submodule with ssh, and if
you cloned with https, the submodule would be cloned with https.