Fixed Broken Matrix Multiplication
The list in the motor message (data) needed to be initialized to all zeroes, and each value from the result of the matrix multiplication must be assigned to that list. This is because numpy's matmul returns an ndarray and not a list, and therefore cannot be directly copied.

TODO: Use the  function on the resultant ndarray to do direct assignment.