Handle rov launch front camera regex return values
It seems that there has been a change in python? and now the re.search
function can, instead of returning an empty list, return None. So, that
has to be handled with a match statement. This also means now that if
the front camera is not connected, it will fail gracefully. As I've
implemented here, it will default to the top camera path. Hacks on top
of hacks.