Added countdown widget
A 'CountdownWidget'
includes a timer which counts down from some amount of time. The timer may be provided with a default
value to start at, this can always be modified using the application at runtime. The widget
also includes a start, pause/resume and stop button. Once the timer reaches 00:00, it begins counting
with negative values until the user terminates the timer