CSS style file for check list widget
Put all CSS for the checklist here and remove any you currently have in the .py file.
For this CSS to apply, add this code to the end of your init

with open(style_sheet_file) as style_sheet:
    self.setStyleSheet(style_sheet.read())

The color constants do not want to import for some reason, ignore that for now, we can add that later.
To fix the issue with the dots, choose some mono spaced font as the font for your check box
(see this file, there is a note where you want to add it) and then it should work B)