Clicking the 0 key changes color mode. #209
Started implemetation of light and dark mode. After some idea iterations, it was decided the best way to go about this was to have a hot key which switches between light and dark mode so we cah choose based on the situation. This key was chosen to be zero by can be easily adjusted. Currently all widgets but the countdown change color upon pressing zero.

TODO:
- Update light mode color constants so they are less disgusting
- Maybe change dark mode accent color to orange
- Create svg assets for both modes and add paths to dictionary (talk to ava)
- Countdown widget color update, its harder than the others because the colors change at runtime already