Parameter set for gazebo republisher
Gazebo requires a singular float value for the newtons of force a thruster should generate. These values have to each be sent on their own unique topics. Previously we used another node which republishes the values from `motor_values` to topics. However a cleaner solution would be to have `thrust` have a boolean parameter which can be set to true when we want to have these additional publishers. This implementation is reflected in this commit